--- tests/tests/with_deps/pos/surface/resolver04.rs
@@ -1,15 +1,0 @@
-// Test that we support `super` in name resolution
-
-struct S;
-
-mod a {
-    use super::*;
-
-    #[flux_rs::spec(fn(S))]
-    fn foo(s: S) {}
-}
-
-mod b {
-    #[flux_rs::spec(fn(super::S))]
-    fn foo(s: super::S) {}
-}
